storage: move some more stuff out of reader/reader
